remove models from previous generations

+++ pie/optimize.py
@@ -129,8 +129,9 @@ def run_optimize(train_fn, settings, opt, n_iter, **kwargs):
     - n_iter: int, number of iterations to run
     """
     HALVING = 3
-    ROUNDS = 3
-    INIT_POP = ROUNDS * HALVING
+    ROUNDS = 4
+    INIT_POP = HALVING ** ROUNDS
+    THRESHOLD_SCORE = 0.1
     model_pool: list[tuple[Any, float, str]] = []
     for _ in range(INIT_POP):
         tmp = sample_from_config(opt)
@@ -154,29 +155,45 @@ def run_optimize(train_fn, settings, opt, n_iter, **kwargs):
             print(yaml.dump(dict(merged)))
             try:
                 model_path, scoring = train_fn(check_settings(merge_task_defaults(merged)), **kwargs)
-                model_pool[i] = (model[0], scoring[0]['all']['accuracy'], model_path)
+                final_score = scoring[0]['all']['accuracy']
+
+                # Remove model from the previous generation
+                if model[2] != "":
+                    if os.path.exists(model[2]):
+                        os.remove(model[2])
+
+                if final_score < THRESHOLD_SCORE:
+                    print(f"::: Score too low, removing model :::")
+                    # remove the model file if the score is too low
+                    # to make space for new models in stead of waiting for the halving
+                    if os.path.exists(model_path):
+                        os.remove(model_path)
+                    model_pool[i] = (model[0], 0, "")
+                else:
+                    model_pool[i] = (model[0], final_score, model_path)
+                
             except Exception as e:
                 print(f"::: Exception in run {i+1} of generation {generation}, continuing anyway :::\n{e}")
         
-        print(f"::: Best score in generation {generation}: {model_pool[0][1]} :::")
 
         if len(model_pool) == 1:
             break
         # sort by score
         model_pool = sorted(model_pool, key=lambda x: x[1], reverse=True)
+        print(f"::: Best score in generation {generation}: {model_pool[0][1]} :::")
         # remove worst half
         to_be_removed = model_pool[len(model_pool) // HALVING:]
         # keep best half
         model_pool = model_pool[:len(model_pool) // HALVING]
         # additionally remove any models from the model_pool whose score near zero
-        THRESHOLD_SCORE = 0.01
         to_be_removed += [model for model in model_pool if model[1] <= THRESHOLD_SCORE]
         model_pool = [model for model in model_pool if model[1] > THRESHOLD_SCORE]        
         # actually remove the models files
         print(f"Removing {len(to_be_removed)} models")
         for model in to_be_removed:
             if model[2] != "": # due to an exception, the model might not have been saved
-                os.remove(model[2])
+                if os.path.exists(model[2]):
+                    os.remove(model[2])
         
         generation += 1
